Moved enabling feature segments to AtbInitializer from StatisticsRequester. Removed internalBuild filter
