+++ app/src/androidTest/java/com/duckduckgo/app/statistics/api/StatisticsRequesterJsonTest.kt
@@ -26,7 +26,6 @@ import com.duckduckgo.app.global.AppUrl.ParamKey
 import com.duckduckgo.app.global.plugins.PluginPoint
 import com.duckduckgo.app.statistics.Variant
 import com.duckduckgo.app.statistics.VariantManager
-import com.duckduckgo.app.statistics.api.featureusage.FeatureSegmentsManager
 import com.duckduckgo.app.statistics.model.Atb
 import com.duckduckgo.app.statistics.store.StatisticsDataStore
 import com.duckduckgo.app.statistics.store.StatisticsSharedPreferences
@@ -58,7 +57,6 @@ class StatisticsRequesterJsonTest {
     private lateinit var statisticsStore: StatisticsDataStore
     private lateinit var testee: StatisticsRequester
     private var mockEmailManager: EmailManager = mock()
-    private val mockFeatureSegmentsManager: FeatureSegmentsManager = mock()
 
     private val server = MockWebServer()
 
@@ -78,7 +76,7 @@ class StatisticsRequesterJsonTest {
                 return listOf()
             }
         }
-        testee = StatisticsRequester(statisticsStore, statisticsService, mockVariantManager, plugins, mockEmailManager, mockFeatureSegmentsManager)
+        testee = StatisticsRequester(statisticsStore, statisticsService, mockVariantManager, plugins, mockEmailManager)
         whenever(mockVariantManager.getVariant()).thenReturn(Variant("ma", 100.0, filterBy = { true }))
     }
