+++ app/src/test/java/com/duckduckgo/app/statistics/AtbInitializerTest.kt
@@ -18,12 +18,14 @@ package com.duckduckgo.app.statistics
 
 import com.duckduckgo.app.referral.StubAppReferrerFoundStateListener
 import com.duckduckgo.app.statistics.api.StatisticsUpdater
+import com.duckduckgo.app.statistics.api.featureusage.FeatureSegmentsManager
 import com.duckduckgo.app.statistics.store.StatisticsDataStore
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.never
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 
@@ -34,13 +36,14 @@ class AtbInitializerTest {
 
     private val statisticsDataStore: StatisticsDataStore = mock()
     private val statisticsUpdater: StatisticsUpdater = mock()
+    private val mockFeatureSegmentsManager: FeatureSegmentsManager = mock()
     private lateinit var appReferrerStateListener: AtbInitializerListener
 
     @Test
     fun whenReferrerInformationInstantlyAvailableThenAtbInitialized() = runTest {
         whenever(statisticsDataStore.hasInstallationStatistics).thenReturn(false)
         appReferrerStateListener = StubAppReferrerFoundStateListener(referrer = "xx")
-        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener))
+        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener), mockFeatureSegmentsManager)
 
         testee.initialize()
 
@@ -51,7 +54,7 @@ class AtbInitializerTest {
     fun whenReferrerInformationQuicklyAvailableThenAtbInitialized() = runTest {
         whenever(statisticsDataStore.hasInstallationStatistics).thenReturn(false)
         appReferrerStateListener = StubAppReferrerFoundStateListener(referrer = "xx", mockDelayMs = 1000L)
-        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener))
+        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener), mockFeatureSegmentsManager)
 
         testee.initialize()
 
@@ -62,7 +65,7 @@ class AtbInitializerTest {
     fun whenReferrerInformationTimesOutThenAtbInitialized() = runTest {
         whenever(statisticsDataStore.hasInstallationStatistics).thenReturn(false)
         appReferrerStateListener = StubAppReferrerFoundStateListener(referrer = "xx", mockDelayMs = Long.MAX_VALUE)
-        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener))
+        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener), mockFeatureSegmentsManager)
 
         testee.initialize()
 
@@ -72,10 +75,27 @@ class AtbInitializerTest {
     @Test
     fun whenAlreadyInitializedThenRefreshCalled() = runTest {
         configureAlreadyInitialized()
-        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener))
+        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener), mockFeatureSegmentsManager)
 
         testee.initialize()
         verify(statisticsUpdater).refreshAppRetentionAtb()
+    }
+
+    @Test
+    fun whenStatisticsNotStoredThenDontEnableFeatureSegment() {
+        whenever(statisticsDataStore.hasInstallationStatistics).thenReturn(false)
+        appReferrerStateListener = StubAppReferrerFoundStateListener(referrer = "xx")
+        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener), mockFeatureSegmentsManager)
+
+        verify(mockFeatureSegmentsManager, never()).enableSendPixelForFeatureSegments()
+    }
+
+    @Test
+    fun whenStatisticsStoredThenDontEnableFeatureSegment() {
+        configureAlreadyInitialized()
+        testee = AtbInitializer(TestScope(), statisticsDataStore, statisticsUpdater, setOf(appReferrerStateListener), mockFeatureSegmentsManager)
+
+        verify(mockFeatureSegmentsManager, never()).enableSendPixelForFeatureSegments()
     }
 
     private fun configureAlreadyInitialized() {
