+++ app/src/test/java/com/duckduckgo/app/statistics/api/StatisticsRequesterTest.kt
@@ -21,7 +21,6 @@ import com.duckduckgo.app.email.EmailManager
 import com.duckduckgo.app.global.plugins.PluginPoint
 import com.duckduckgo.app.statistics.Variant
 import com.duckduckgo.app.statistics.VariantManager
-import com.duckduckgo.app.statistics.api.featureusage.FeatureSegmentsManager
 import com.duckduckgo.app.statistics.model.Atb
 import com.duckduckgo.app.statistics.store.StatisticsDataStore
 import io.reactivex.Observable
@@ -38,7 +37,6 @@ class StatisticsRequesterTest {
     private var mockResponseBody: ResponseBody = mock()
     private var mockVariantManager: VariantManager = mock()
     private val mockEmailManager: EmailManager = mock()
-    private val mockFeatureSegmentsManager: FeatureSegmentsManager = mock()
 
     private val plugins = object : PluginPoint<RefreshRetentionAtbPlugin> {
         override fun getPlugins(): Collection<RefreshRetentionAtbPlugin> {
@@ -46,7 +44,7 @@ class StatisticsRequesterTest {
         }
     }
     private var testee: StatisticsRequester =
-        StatisticsRequester(mockStatisticsStore, mockService, mockVariantManager, plugins, mockEmailManager, mockFeatureSegmentsManager)
+        StatisticsRequester(mockStatisticsStore, mockService, mockVariantManager, plugins, mockEmailManager)
 
     @get:Rule
     @Suppress("unused")
@@ -176,20 +174,6 @@ class StatisticsRequesterTest {
         testee.refreshAppRetentionAtb()
 
         verify(mockService).updateAppAtb(any(), any(), any(), eq(1))
-    }
-
-    @Test
-    fun whenNoStatisticsStoredThenEnableFeatureSegment() {
-        configureNoStoredStatistics()
-        testee.initializeAtb()
-        verify(mockFeatureSegmentsManager).enableSendPixelForFeatureSegments()
-    }
-
-    @Test
-    fun whenStatisticsStoredThenDontEnableFeatureSegment() {
-        configureStoredStatistics()
-        testee.initializeAtb()
-        verify(mockFeatureSegmentsManager, never()).enableSendPixelForFeatureSegments()
     }
 
     private fun configureNoStoredStatistics() {
